backend/requirements.txt
# Django框架相关
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
drf-spectacular==0.26.5

# 数据库相关
psycopg2-binary==2.9.7
django-redis==5.4.0

# 异步任务处理
celery==5.3.4
redis==5.0.1

# WebSocket支持
channels==4.0.0
channels-redis==4.1.0

# 配置管理
python-decouple==3.8

# HTTP请求
requests==2.31.0
httpx==0.25.1
aiohttp==3.9.1

# 智能体和AI相关
langchain==0.0.330
openai==1.3.3
anthropic==0.3.11
sentence-transformers==2.2.2

# 向量数据库
faiss-cpu==1.7.4
chromadb==0.4.15

# 自然语言处理
jieba==0.42.1
pypinyin==0.49.0

# 数据处理
pandas==2.1.3
numpy==1.24.3

# 文档处理
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-magic==0.4.27
chardet==5.2.0

# 日期时间处理
python-dateutil==2.8.2

# 密码学
cryptography==41.0.7

# 媒体文件处理
Pillow==10.1.0

# 文档生成
Markdown==3.5.1

# 数据验证
pydantic==2.5.0

# 测试相关
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# 代码格式化
black==23.10.1
isort==5.12.0

# 开发工具
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# 监控和日志
sentry-sdk==1.38.0

# 部署相关
gunicorn==21.2.0
whitenoise==6.6.0 